so_long done ? just normed to do

+++ error.c
@@ -6,13 +6,13 @@
 /*   By: mleschev <mleschev@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/01 15:37:49 by mleschev          #+#    #+#             */
-/*   Updated: 2025/05/22 23:46:51 by mleschev         ###   ########.fr       */
+/*   Updated: 2025/05/27 16:12:49 by mleschev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
 
-void	check_map_error(mlx_window *params)
+void	check_map_error(t_mlx_window *params)
 {
 	if (params->p_nbr == 0)
 		print_error("No spawn point", params);
@@ -24,17 +24,25 @@ void	check_map_error(mlx_window *params)
 		print_error("Need at least one collectible", params);
 }
 
-void	print_error(char *error, mlx_window *params)
+void	print_error(char *error, t_mlx_window *params)
 {
-	kill_process(params);
+	if (params)
+		kill_process(params);
 	ft_printf("Error\n%s", error);
 	exit (-1);
 }
-void	kill_process(mlx_window *map)
+
+void	kill_process(t_mlx_window *map)
 {
+	close(map->fd);
+	open(map->path, O_RDONLY);
 	if (map->buf)
 	{
 		free(map->buf);
-		map->buf = NULL;
+	}
+	while (map->buf)
+	{
+		map->buf = get_next_line(map->fd);
+		free(map->buf);
 	}
 }
